Feature: Made the RecyclerView items deletable by swiping left or right
I use Android's ItemTouchHelper and its abstract class SimpleCallback to implement the swipe behaviour. When an item is swiped to the right or to the left, I store the item in a variable, delete it from the database, refresh the UI and render an error Snackbar with an UNDO button. When the UNDO button is clicked, I re-insert the item from the variable back into the database and refresh the UI again.

DebugException.kt is used to test try catch clauses.